Feat/cluster mode

+++ index.js
@@ -1,11 +1,45 @@
-import {callbackify} from 'node:util';
-import {createClient} from 'redis';
+import {createClient, createCluster} from 'redis';
 
 export async function redisStore(config) {
   const redisCache = createClient(config);
   await redisCache.connect();
 
   return buildRedisStoreWithConfig(redisCache, config);
+}
+
+export async function redisClusterStore(config) {
+  const redisCache = createCluster(config)
+  await redisCache.connect()
+
+  return buildRedisStoreWithConfig(redisCache, config)
+}
+
+export async function redisAdaptativeConnection(...configs) {
+  for(let configIndex = 0; configIndex < configs.length; configIndex++) {
+    const config = configs[configIndex]
+    if(config.socket) {
+      try {
+        return {
+          mode: 'master/slave',
+          connection: await redisStore(config)
+        }
+      } catch (e) {
+        console.info(`Could not connect master/slave with configuration at index:[${configIndex}]`)
+      }
+    } else if(config.rootNodes) {
+      try {
+        return {
+          mode: 'cluster',
+          connection: await redisClusterStore(config)
+        }
+      } catch (e) {
+        console.info(`Could not connect Cluster with configuration at index:[${configIndex}]`)
+      }
+    } else {
+      throw 'Missing socket or rootNodes field'
+    }
+  }
+  throw 'None of the configurations lead to a proper connection'
 }
 
 const buildRedisStoreWithConfig = (redisCache, config) => {
@@ -97,9 +131,17 @@ const buildRedisStoreWithConfig = (redisCache, config) => {
     return redisCache.del(args);
   };
   const reset = async () => {
+    const couldFlush = redisCache.flushDb
+    if(!couldFlush) {
+      throw 'Cannot Reset'
+    }
     return redisCache.flushDb();
   };
   const keys = async (pattern) => {
+    const hasKeys = redisCache.keys
+    if(!hasKeys) {
+      throw 'Has No Keys Function'
+    }
     return redisCache.keys(pattern);
   };
   const ttl = async (key) => {
@@ -110,18 +152,10 @@ const buildRedisStoreWithConfig = (redisCache, config) => {
     name: 'redis',
     getClient: () => redisCache,
     isCacheableValue,
-    set: (key, value, options, cb) => {
-      if (typeof options === 'function') {
-        cb = options;
-        options = {};
-      }
+    set: (key, value, options) => {
       options = options || {};
 
-      if (typeof cb === 'function') {
-        callbackify(set)(key, value, options, cb);
-      } else {
-        return set(key, value, options);
-      }
+      return set(key, value, options);
     },
     get: (key, options, cb) => {
       if (typeof options === 'function') {
@@ -130,64 +164,26 @@ const buildRedisStoreWithConfig = (redisCache, config) => {
       }
       options = options || {};
 
-      if (typeof cb === 'function') {
-        callbackify(get)(key, options, cb);
-      } else {
-        return get(key, options);
-      }
+      return get(key, options);
     },
     del: (...args) => {
-      if (typeof args.at(-1) === 'function') {
-        const cb = args.pop();
-        callbackify(del)(args, cb);
-      } else {
-        return del(args);
-      }
+      return del(args);
     },
     mset: (...args) => {
-      if (typeof args.at(-1) === 'function') {
-        const cb = args.pop();
-        callbackify(mset)(args, cb);
-      } else {
-        return mset(args);
-      }
+      return mset(args);
     },
     mget: (...args) => {
-      if (typeof args.at(-1) === 'function') {
-        const cb = args.pop();
-        callbackify(() => mget(...args))(cb);
-      } else {
-        return mget(...args);
-      }
+      return mget(...args);
     },
     mdel: (...args) => {
-      if (typeof args.at(-1) === 'function') {
-        const cb = args.pop();
-        callbackify(() => mdel(...args))(cb);
-      } else {
-        return mdel(...args);
-      }
+      return mdel(...args);
     },
-    reset: (cb) => {
-      if (typeof cb === 'function') {
-        callbackify(reset)(cb);
-      } else {
-        return reset();
-      }
-    },
+    reset,
     keys: (pattern = '*', cb) => {
-      if (typeof cb === 'function') {
-        callbackify(() => keys(pattern))(cb);
-      } else {
-        return keys(pattern);
-      }
+      return keys(pattern);
     },
-    ttl: (key, cb) => {
-      if (typeof cb === 'function') {
-        callbackify(() => ttl(key))(cb);
-      } else {
-        return ttl(key);
-      }
+    ttl: (key) => {
+      return ttl(key);
     },
   };
 };
